Refactor : SharedScreen 관련 파일들 위치 변경
컴포넌트 내로 관련 코드들을 이동하였습니다.
