+++ frontend/src/components/Cam/MainScreen.tsx
@@ -2,12 +2,12 @@ import React, { useContext } from 'react';
 import styled from 'styled-components';
 
 import { ToggleStoreContext } from './ToggleStore';
-import SharedScreen from './SharedScreen';
+import SharedScreen from './SharedScreen/SharedScreen';
 import { CamStoreContext } from './CamStore';
 import type { Screen } from '../../types/cam';
 import LocalUserScreen from './LocalUserScreen';
 import UserScreen from './UserScreen';
-import { SharedScreenStoreContext } from './SharedScreenStore';
+import { SharedScreenStoreContext } from './SharedScreen/SharedScreenStore';
 
 const Container = styled.div<{ activeTab: string[]; isMouseOnCamPage: boolean; numOfScreen: number }>`
   width: ${(props) => props.activeTab[0]};
